make export_data_as_npz take data type as paraeter

+++ src/data/export_numpy.py
@@ -2,7 +2,8 @@ import numpy as np
 import torch
 
 
-def export_data_as_npz(path, data_loader, num_batches=None, feature_transform=None, label_transform=None, seed=None):
+def export_data_as_npz(path, data_loader, num_batches=None, feature_transform=None, label_transform=None,
+                       dtype='int8', seed=None):
     np_features = np.empty(0)
     np_labels = np.empty(0)
 
@@ -21,6 +22,7 @@ def export_data_as_npz(path, data_loader, num_batches=None, feature_transform=No
         np_features = np.concatenate((np_features, features)) if np_features.shape[0] else features
         np_labels = np.concatenate((np_labels, labels)) if np_labels.shape[0] else labels
 
+    np_features = np_features.astype(dtype)
     np.savez(path, features=np_features, labels=np_labels)
